+++ packages/components/src/components/table/table/table.tsx
@@ -14,6 +14,7 @@ const propTypes: PropTypes<typeof Table> = {
   caption: AllowedTypes.string,
   compact: AllowedTypes.boolean,
   layout: AllowedTypes.oneOf<TableLayout>(TABLE_LAYOUTS),
+  sticky: AllowedTypes.boolean,
 };
 
 /**
@@ -31,11 +32,16 @@ export class Table {
    * Use an element with an attribute of `slot="caption"` for a visible caption. */
   @Prop() public caption?: string;
 
-  /** Displays as compact version. */
+  /** Displays with reduced spacing and smaller padding for a more condensed layout. */
   @Prop() public compact?: boolean = false;
 
   /** Controls the layout behavior of the table. */
   @Prop() public layout?: TableLayout = 'auto';
+
+  /**
+   * @experimental Makes the scroll indicator sticky at the top or bottom while scrolling depending on the scroll direction.
+   */
+  @Prop() public sticky?: boolean = false;
 
   /** Emitted when sorting is changed. */
   @Event({ bubbles: false }) public update: EventEmitter<TableUpdateEventDetail>;
@@ -66,8 +72,7 @@ export class Table {
             <slot name="caption" />
           </div>
         )}
-
-        <PrefixedTagNames.pScroller scrollbar={true}>
+        <PrefixedTagNames.pScroller scrollbar={true} compact={this.compact} sticky={this.sticky}>
           <div class="table" role="table" {...tableAttr}>
             <slot />
           </div>