+++ packages/components/src/styles/css-variables/index.ts
@@ -1,4 +1,7 @@
+export * from './blur';
 export * from './border';
 export * from './color';
 export * from './font';
+export * from './motion';
+export * from './shadow';
 export * from './spacing';