+++ packages/components/src/styles/css-variables/motion/duration/durationXl.ts
@@ -1,0 +1,1 @@
+export const durationXl = 'var(--p-duration-xl)' as const;