+++ packages/components/CHANGELOG.md
@@ -16,6 +16,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### Added
 
+- `Scroller`:
+  - Prop `compact`: Displays with reduced spacing and smaller padding for a more condensed layout.
+    ([#4228](https://github.com/porsche-design-system/porsche-design-system/pull/4228))
+  - Prop `sticky`: Makes the indicator sticky at the top or bottom while scrolling depending on the scroll direction.
+    ([#4228](https://github.com/porsche-design-system/porsche-design-system/pull/4228))
+  - CSS Variable `--p-scroller-indicator-top`: Defines the distance from the top of the viewport at which the indicator
+    sticks when scrolling down and `sticky` is enabled.
+    ([#4228](https://github.com/porsche-design-system/porsche-design-system/pull/4228))
+  - CSS Variable `--p-scroller-indicator-bottom`: Defines the distance from the bottom of the viewport at which the
+    indicator sticks when scrolling up and `sticky` is enabled.
+    ([#4228](https://github.com/porsche-design-system/porsche-design-system/pull/4228))
+  - CSS Variable `--p-scroller-gap`: Defines the gap between slotted nodes.
+    ([#4228](https://github.com/porsche-design-system/porsche-design-system/pull/4228))
 - `Tabs Bar`:
   - Prop `background`: Defines the background color. Use `frosted` only on images, videos or gradients.
     ([#4228](https://github.com/porsche-design-system/porsche-design-system/pull/4228))
@@ -53,11 +66,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### Deprecated
 
+- `Scroller`:
+  - Prop `alignScrollIndicator`: Will be removed with next major release, has no effect anymore.
+    ([#4228](https://github.com/porsche-design-system/porsche-design-system/pull/4228))
+  - Prop `scrollToPosition`: Will be removed with next major release, use native `scrollIntoView()` on the slotted
+    element itself, e.g.
+    `el.scrollIntoView({ behavior: 'smooth', block: 'nearest', inline: 'center', container: 'nearest' })`.
+    ([#4228](https://github.com/porsche-design-system/porsche-design-system/pull/4228))
 - `Tabs Bar`:
-  - Prop `weight`: Has no effect anymore.
+  - Prop `weight`: Will be removed with next major release, has no effect anymore.
     ([#4228](https://github.com/porsche-design-system/porsche-design-system/pull/4228))
 - `Tabs`:
-  - Prop `weight`: Has no effect anymore.
+  - Prop `weight`: Will be removed with next major release, has no effect anymore.
     ([#4228](https://github.com/porsche-design-system/porsche-design-system/pull/4228))
 
 ## [4.0.0-beta.1] - 2026-03-02